Created using Colab
Task Overview with 'View Only' Links
You want to create a Python program that:

Accesses an Online Folder: Get a link to a folder that has pictures of people.
Gets Permissions: Make sure the program has permission to see and change the files in that folder.
Recognizes Headshots: Look at each picture and find the paper with a special code that each person is holding.
Groups Photos: Put all the pictures of the same person into a new folder that is named after their code.
Confirm Completion: Tell the user when it's done and ask them to check everything.
Create a CSV File: If the user wants, create a list of all the folders with their codes in a CSV file for later use, ensuring the links are set to 'View Only'.
Step-by-Step Breakdown
Accessing the Folder

You will ask the user to input the link to the online folder.
The program will check if it can access the folder (like asking for permission).
Finding Headshots

You will loop through each picture in the folder.
For each picture, you will look for the paper with a code on it.
Grouping Photos

Once you have the code, you will create a new folder named after that code.
You will move all pictures of that person into this new folder.
User Confirmation

After moving the pictures, the program will let the user know it's done.
It will ask the user if they want to check the folders.
Creating a CSV File

If the user says "yes", the program needs to:
Gather the links to each folder.
Set Each Link to 'View Only': This means that anyone who gets the link can see the pictures, but they cannot delete or change them.
Add their respective codes to another column in the CSV file.
The CSV file will be saved in a folder called "Talent CSV".

What Will You Need?
Libraries like os (to work with files and folders).
A library for handling images and recognizing text (like opencv or pytesseract).
A library for handling CSV files (like csv or pandas).
A way to access online folders, which could involve using APIs (like Google Drive API).
You will need to use the API to change the permissions of the links to 'View Only'.
Summary
This program is like a helper that organizes pictures based on codes while ensuring that the links to the pictures are secure. It starts by asking for a link, checks permissions, finds photos, groups them, confirms with the user, and finally creates a list of links (set to 'View Only') and codes in a CSV file.